+++ src/py/mat3ra/made/tools/utils/__init__.py
@@ -5,6 +5,7 @@ import numpy as np
 from mat3ra.made.material import Material
 from mat3ra.made.utils import ArrayWithIds
 from mat3ra.utils.matrix import convert_2x2_to_3x3
+from scipy.spatial import cKDTree
 
 from ..third_party import PymatgenStructure
 from .coordinate import (
@@ -62,6 +63,23 @@ def get_distance_between_coordinates(coordinate1: List[float], coordinate2: List
         float: The distance between the two coordinates.
     """
     return float(np.linalg.norm(np.array(coordinate1) - np.array(coordinate2)))
+
+
+def calculate_norm_of_distances_between_coordinates(coords1: np.ndarray, coords2: np.ndarray) -> float:
+    """
+    Calculate the norm of distances between two sets of coordinates.
+
+    Args:
+        coords1 (np.ndarray): The first set of coordinates.
+        coords2 (np.ndarray): The second set of coordinates.
+
+    Returns:
+        float: The calculated norm.
+    """
+    tree = cKDTree(coords2)
+    distances, _ = tree.query(coords1)
+    distances = distances[~np.isinf(distances)]
+    return float(np.linalg.norm(distances))
 
 
 def get_norm(vector: List[float]) -> float:
